fixed_seeders

+++ app/Http/Controllers/ClasseController.php
@@ -2,10 +2,11 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\ClassRequest;
 use App\Models\Classe;
+use App\Models\Option;
 use App\Models\School;
 use App\Services\ClassService;
+use App\Http\Requests\ClassRequest;
 use Illuminate\Http\Request;
 
 class ClasseController extends Controller
@@ -15,61 +16,83 @@ class ClasseController extends Controller
     public function __construct(ClassService $classService)
     {
         $this->classService = $classService;
-        $this->middleware('auth');
-        // Retirez temporairement cette ligne pour déboguer
-        // $this->authorizeResource(Classes::class, 'class');
     }
 
-    // Ajoutez les autorisations manuellement dans chaque méthode
-    public function index()
+    public function index(Request $request)
     {
-        $currentUser = auth()->user();
-        $classes = $this->classService->getClassesList($currentUser, [
-            'name' => request('name'),
-            'school' => request('school')
+        $this->authorize('viewAny', Classe::class);
+
+        $classes = $this->classService->getClassesList(auth()->user(), [
+            'name' => $request->get('name'),
+            'school' => $request->get('school')
         ]);
 
         return view('classes.index', compact('classes'));
     }
 
-
     public function create()
     {
         $this->authorize('create', Classe::class);
-        // Si c'est un admin, on ne montre que son école
-        if (auth()->user()->canManageOwnClasses()) {
-            $schools = School::where('id', auth()->user()->school_id)->get();
-        } // Si c'est un super admin, on montre toutes les écoles
-        else {
-            $schools = School::all();
-        }
-        return view('classes.form', compact('schools'));
+
+        $options = Option::when(auth()->user()->role->name !== 'Super Administrateur', function($query) {
+            return $query->where('school_id', auth()->user()->school_id);
+        })->get();
+
+        // Récupérer les écoles selon le rôle
+        $schools = auth()->user()->role->name === 'Super Administrateur'
+            ? School::all()
+            : School::where('id', auth()->user()->school_id)->get();
+
+        return view('classes.form', compact('options', 'schools'));
     }
+
 
     public function store(ClassRequest $request)
     {
+        $this->authorize('create', Classe::class);
+
         $class = $this->classService->store($request->validated(), auth()->user());
-        return redirect()->route('classes.index')
-            ->with('success', 'Classe créée avec succès.');
+
+        return redirect()
+            ->route('classes.index')
+            ->with('success', 'Classe créée avec succès');
     }
 
     public function edit(Classe $class)
     {
-        $schools = School::all();
-        return view('classes.form', compact('class', 'schools'));
+        $this->authorize('update', $class);
+
+        $options = Option::when(auth()->user()->role->name !== 'Super Administrateur', function($query) {
+            return $query->where('school_id', auth()->user()->school_id);
+        })->get();
+
+        // Récupérer les écoles selon le rôle
+        $schools = auth()->user()->role->name === 'Super Administrateur'
+            ? School::all()
+            : School::where('id', auth()->user()->school_id)->get();
+
+        return view('classes.form', compact('class', 'options', 'schools'));
     }
 
     public function update(ClassRequest $request, Classe $class)
     {
+        $this->authorize('update', $class);
+
         $this->classService->update($class, $request->validated());
-        return redirect()->route('classes.index')
-            ->with('success', 'Classe mise à jour avec succès.');
+
+        return redirect()
+            ->route('classes.index')
+            ->with('success', 'Classe mise à jour avec succès');
     }
 
     public function destroy(Classe $class)
     {
+        $this->authorize('delete', $class);
+
         $this->classService->delete($class);
-        return redirect()->route('classes.index')
-            ->with('success', 'Classe supprimée avec succès.');
+
+        return redirect()
+            ->route('classes.index')
+            ->with('success', 'Classe supprimée avec succès');
     }
 }
